refactor: scan.rs in render_implementations and removed editor examples
